+++ libs/dot-components/src/lib/components/link/Link.styles.ts
@@ -1,0 +1,14 @@
+import { Link } from '@material-ui/core';
+import styled from 'styled-components';
+
+export const rootClassName = 'dot-link';
+
+export const StyledLink = styled(Link)`
+  &.${rootClassName} {
+    cursor: pointer;
+
+    &:hover :not(.MuiLink-underlineHover) {
+      text-decoration: none;
+    }
+  }
+`;